refactor k8s code to java flavor
