+++ server/controller/src/main/java/ai/starwhale/mlops/schedule/SWTaskScheduler.java
@@ -41,11 +41,4 @@ public interface SWTaskScheduler {
      */
     void stopSchedule(Collection<Long> taskIds);
 
-    /**
-     * pop tasks available to the node. if no task is available or the node is full, an empty list should be returned
-     * TaskStatus -> Created.DOING
-     * @param node the node load info
-     * @return tasks to be assigned to the node
-     */
-    List<Task> schedule(Node node);
 }
