+++ server/controller/src/main/resources/application.yaml
@@ -17,6 +17,9 @@ sw:
     issuer: ${SW_JWT_ISSUER:starWhale}
     expireMinutes: ${SW_JWT_TOKEN_EXPIRE_MINUTES:43200}
   taskSize: ${SW_TASK_SPLIT_SIZE:256}
+  infra:
+    k8s:
+      name-space: ${SW_K8S_NAME_SPACE:default}
   storage:
     path-prefix: ${SW_STORAGE_PREFIX:StarWhale}
     s3-config:
